+++ lib/widgets/habits/habit_streaks.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:lotti/blocs/habits/habits_cubit.dart';
 import 'package:lotti/blocs/habits/habits_state.dart';
-import 'package:lotti/themes/theme.dart';
+import 'package:lotti/widgets/charts/habits/habit_completion_rate_chart.dart';
 
 class HabitStreaksCounter extends StatelessWidget {
   const HabitStreaksCounter({super.key});
@@ -16,22 +16,16 @@ class HabitStreaksCounter extends StatelessWidget {
 
         return Column(
           children: [
-            Text(
-              '$total habits total',
-              style: chartTitleStyle(),
-            ),
-            Text(
-              '$todayCount completed today',
-              style: chartTitleStyle(),
-            ),
-            Text(
-              '${state.shortStreakCount} short streaks of 3+ days',
-              style: chartTitleStyle(),
-            ),
-            Text(
-              '${state.longStreakCount} long streaks of 7+ days',
-              style: chartTitleStyle(),
-            ),
+            InfoLabel('$todayCount out of $total habits completed today'),
+            // TODO: bring back display of streaks
+            // Text(
+            //   '${state.shortStreakCount} short streaks of 3+ days',
+            //   style: chartTitleStyle(),
+            // ),
+            // Text(
+            //   '${state.longStreakCount} long streaks of 7+ days',
+            //   style: chartTitleStyle(),
+            // ),
           ],
         );
       },