+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - FormBuilderCupertinoDateTimePicker replaced in new measurement dialog
 - New measurement page removed
 - flutter_datetime_picker removed from entry datetime modal
+- Simpler display of habit completion count
 
 ## [0.8.291] - 2023-03-27
 ### Fixed: